Implement some more things with DamageEvents. Push the initial handling to CraftBukkit 1.7.10 compatibility.

+++ src/main/java/com/afterkraft/kraftrpg/skills/ActiveSkillRunner.java
@@ -215,7 +215,7 @@ public final class ActiveSkillRunner {
         if (result == SkillCastResult.NORMAL) {
             caster.setHealth(caster.getHealth() - healthCost);
             caster.setMana(caster.getMana() - (int) manaCost);
-            caster.modifyStamina((float) -hungerCost);
+            caster.modifyStamina((int) -hungerCost);
             caster.getInventory().removeItem(reagent);
 
             double exp = plugin.getSkillConfigManager().getUsedDoubleSetting(caster, skill, SkillSetting.EXP_ON_CAST);
